Move HttpError to http-error
