+++ src/api/client.ts
@@ -1,27 +1,6 @@
 import { YearQuarter } from '../year-quarter'
 import { UrlBuilder } from './url-builder'
-
-export class HttpError implements Error {
-  public name = 'HttpError'
-  public message: string
-
-  // eslint-disable-next-line @typescript-eslint/camelcase
-  constructor(public response: GoogleAppsScript.URL_Fetch.HTTPResponse) {
-    this.message = `${response.getResponseCode()}: ${response.getContentText()}`
-  }
-
-  public isInvalidTestingRequest(): boolean {
-    const content = this.response.getContentText()
-    return (
-      content ===
-      '{"message":"Testing Apikey is only allowed to ticker ending with \\"01\\""}'
-    )
-  }
-
-  public toString(): string {
-    return this.message
-  }
-}
+import { HttpError } from './http-error'
 
 export class BuffettCodeApiClientV2 {
   static readonly baseUrl = 'https://api.buffett-code.com/api/v2'
