+++ src/custom-functions/bcode.ts
@@ -1,7 +1,8 @@
 import { ApiResponseError } from './error'
 import { bcodeIndicator } from './bcode-indicator'
 import { bcodeQuarter } from './bcode-quarter'
-import { BuffettCodeApiClientV2, HttpError } from '../api/client'
+import { BuffettCodeApiClientV2 } from '../api/client'
+import { HttpError } from '../api/http-error'
 import { IndicatorProperty } from '../api/indicator-property'
 import { QuarterProperty } from '../api/quarter-property'
 import { Result } from '../result'
